modify null check & notify
